fix(tui): compose tree prefixes per level; restore domain→service and (no aliases)
Review fixes for the ASCII topology rendering:
- Service-row prefixes now compose the network-level continuation column
  with the service-level branch connector instead of selecting between
  them — previously services under the last network (the common
  single-network case) rendered with no tree structure at all.
- Restore the domain → service mapping in the Public edge section.
- Restore the yellow '(no aliases)' indicator, which had become dead code.
- Warning rows keep both continuation columns so ⚠ aligns under the
  service name.

Split networks.rs into networks/{view,render} to stay under the 250-line
cap, and add tree-shape tests that lock the prefix geometry.
